+++ packages/firebase-dart/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.4.0
+
+* Updated for Firebase api v2.2.2
+
+* Deprecated `name` getter on Firebase and DataSnapshot
+
+* Added `key` getter on Firebase and DataSnapshot, replacing `name`
+
+* Added changeEmail()
+
+* Added authAnonymously(), authWithOAuthPopup(), authWithOAuthRedirect()
+
+* Added getAuth() and onAuth() listener
+
+* Added orderByChild(), orderByKey(), orderByValue(), orderByPriority()
+
+* Added equalTo(), limitToFirst(), limitToLast()
+
+* Deprecated `limit` on Query object
+
+* Added `exists` getter to DataSnapshot
+
+* Added several tests
+
 ## 0.3.0
 
 * Add createUser(), removeUser() and authWithPassword()